First Attempt to Include Dictionary Completions
Doesn't appear to be working yet, but the goal is to include dictionary
results when using the tab completion feature. This would be ideal, so
that typing longer words could be sped up...also, I'm fairly dyslexic,
and routinely misspell words and this would make it easier to correct
mistakes.